requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
Pillow==10.0.1
pytz==2025.2
sqlparse==0.5.3
